rename BranchesCurrentManager
