+++ src/main/java/com/powsybl/openloadflow/OpenLoadFlowProvider.java
@@ -22,7 +22,7 @@ import com.powsybl.openloadflow.ac.DistributedSlackOuterLoop;
 import com.powsybl.openloadflow.ac.PhaseControlOuterLoop;
 import com.powsybl.openloadflow.ac.ReactiveLimitsOuterLoop;
 import com.powsybl.openloadflow.ac.TransformerVoltageControlOuterLoop;
-import com.powsybl.openloadflow.util.currentBranchesManager.AllBranchesManager;
+import com.powsybl.openloadflow.util.branchesCurrentManager.AllBranchesCurrentManager;
 import com.powsybl.openloadflow.ac.nr.DcValueVoltageInitializer;
 import com.powsybl.openloadflow.ac.nr.DefaultNewtonRaphsonStoppingCriteria;
 import com.powsybl.openloadflow.ac.nr.NewtonRaphsonStatus;
@@ -44,7 +44,7 @@ import com.powsybl.openloadflow.network.impl.Networks;
 import com.powsybl.openloadflow.network.util.ActivePowerDistribution;
 import com.powsybl.openloadflow.util.Markers;
 import com.powsybl.openloadflow.util.PowsyblOpenLoadFlowVersion;
-import com.powsybl.openloadflow.util.currentBranchesManager.CurrentBranchesManager;
+import com.powsybl.openloadflow.util.branchesCurrentManager.BranchesCurrentManager;
 import com.powsybl.tools.PowsyblCoreVersion;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -122,12 +122,12 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
 
     public static AcLoadFlowParameters createAcParameters(Network network, MatrixFactory matrixFactory, LoadFlowParameters parameters,
                                                           OpenLoadFlowParameters parametersExt, boolean breakers) {
-        return createAcParameters(network, matrixFactory, parameters, parametersExt, breakers, false, new AllBranchesManager());
+        return createAcParameters(network, matrixFactory, parameters, parametersExt, breakers, false, new AllBranchesCurrentManager());
     }
 
     public static AcLoadFlowParameters createAcParameters(Network network, MatrixFactory matrixFactory, LoadFlowParameters parameters,
                                                           OpenLoadFlowParameters parametersExt, boolean breakers, boolean forceA1Var,
-                                                          CurrentBranchesManager currentBranchesManager) {
+                                                          BranchesCurrentManager branchesCurrentManager) {
 
         SlackBusSelector slackBusSelector = getSlackBusSelector(network, parameters, parametersExt);
 
@@ -177,7 +177,7 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
                                         parametersExt.getPlausibleActivePowerLimit(),
                                         forceA1Var,
                                         parametersExt.isAddRatioToLinesWithDifferentNominalVoltageAtBothEnds(),
-                                        currentBranchesManager);
+            branchesCurrentManager);
     }
 
     private LoadFlowResult runAc(Network network, LoadFlowParameters parameters, OpenLoadFlowParameters parametersExt) {
