+++ src/main/java/com/powsybl/openloadflow/ac/equations/AcEquationSystemCreationParameters.java
@@ -6,8 +6,8 @@
  */
 package com.powsybl.openloadflow.ac.equations;
 
-import com.powsybl.openloadflow.util.currentBranchesManager.AllBranchesManager;
-import com.powsybl.openloadflow.util.currentBranchesManager.CurrentBranchesManager;
+import com.powsybl.openloadflow.util.branchesCurrentManager.AllBranchesCurrentManager;
+import com.powsybl.openloadflow.util.branchesCurrentManager.BranchesCurrentManager;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -20,21 +20,21 @@ public class AcEquationSystemCreationParameters {
 
     private final boolean forceA1Var;
 
-    private final CurrentBranchesManager currentBranchesManager;
+    private final BranchesCurrentManager branchesCurrentManager;
 
     public AcEquationSystemCreationParameters(boolean phaseControl, boolean transformerVoltageControl) {
         this(phaseControl, transformerVoltageControl, false);
     }
 
     public AcEquationSystemCreationParameters(boolean phaseControl, boolean transformerVoltageControl, boolean forceA1Var) {
-        this(phaseControl, transformerVoltageControl, forceA1Var, new AllBranchesManager());
+        this(phaseControl, transformerVoltageControl, forceA1Var, new AllBranchesCurrentManager());
     }
 
-    public AcEquationSystemCreationParameters(boolean phaseControl, boolean transformerVoltageControl, boolean forceA1Var, CurrentBranchesManager currentBranchesManager) {
+    public AcEquationSystemCreationParameters(boolean phaseControl, boolean transformerVoltageControl, boolean forceA1Var, BranchesCurrentManager branchesCurrentManager) {
         this.phaseControl = phaseControl;
         this.transformerVoltageControl = transformerVoltageControl;
         this.forceA1Var = forceA1Var;
-        this.currentBranchesManager = currentBranchesManager;
+        this.branchesCurrentManager = branchesCurrentManager;
     }
 
     public boolean isPhaseControl() {
@@ -49,7 +49,7 @@ public class AcEquationSystemCreationParameters {
         return forceA1Var;
     }
 
-    public CurrentBranchesManager getCurrentBranchesManager() {
-        return currentBranchesManager;
+    public BranchesCurrentManager getBranchesCurrentManager() {
+        return branchesCurrentManager;
     }
 }
