+++ src/main/java/com/powsybl/openloadflow/util/branchesCurrentManager/AllBranchesCurrentManager.java
@@ -4,23 +4,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.util.currentBranchesManager;
+package com.powsybl.openloadflow.util.branchesCurrentManager;
 
 import com.powsybl.openloadflow.network.LfBranch;
-
-import java.util.Collection;
 
 /**
  * @author Gael Macherel <gael.macherel at artelys.com>
  */
-public class SpecificBranchesManager implements CurrentBranchesManager {
-    private final Collection<LfBranch> branches;
-
-    public SpecificBranchesManager(Collection<LfBranch> branches) {
-        this.branches = branches;
-    }
-
+public class AllBranchesCurrentManager implements BranchesCurrentManager {
     public boolean shouldCreate(LfBranch lfBranch) {
-        return branches.contains(lfBranch);
+        return true;
     }
 }
