+++ src/main/java/com/powsybl/openloadflow/util/branchesCurrentManager/SpecificBranchesCurrentManager.java
@@ -4,15 +4,23 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.util.currentBranchesManager;
+package com.powsybl.openloadflow.util.branchesCurrentManager;
 
 import com.powsybl.openloadflow.network.LfBranch;
+
+import java.util.Collection;
 
 /**
  * @author Gael Macherel <gael.macherel at artelys.com>
  */
-public class AllBranchesManager implements CurrentBranchesManager {
+public class SpecificBranchesCurrentManager implements BranchesCurrentManager {
+    private final Collection<LfBranch> branches;
+
+    public SpecificBranchesCurrentManager(Collection<LfBranch> branches) {
+        this.branches = branches;
+    }
+
     public boolean shouldCreate(LfBranch lfBranch) {
-        return true;
+        return branches.contains(lfBranch);
     }
 }
